# docs/swagger.yaml
definitions:
...
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      lastModified:
        type: string
      name:
        type: string
      profilePhoto:
        type: string
      role:
        type: string
      verified:
        type: boolean
    type: object
  models.Workspace:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      lastModified:
        type: string
      name:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
info:
  contact:
    email: support@hazel.local
    name: API Support
    url: https://github.com/luckmade/hazel
  description: This is the backend API for the Hazel project management application.
  title: Hazel Project Management API
  version: "1.0"
paths:
...
  /workspaces:
    post:
      consumes:
      - application/json
      description: Create a new workspace
      parameters:
      - description: Workspace info
        in: body
        name: workspace
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Workspace'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create workspace
      tags:
      - workspaces
  /workspaces/{id}:
    delete:
      description: Delete a workspace by ID
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete workspace
      tags:
      - workspaces
    get:
      description: Get a workspace by ID
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Workspace'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get workspace
      tags:
      - workspaces
    patch:
      consumes:
      - application/json
      description: Update workspace details
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Workspace update info
        in: body
        name: workspace
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Workspace'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update workspace
      tags:
      - workspaces
  /workspaces/{id}/members:
    get:
      description: Get all members of a workspace
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get workspace members
      tags:
      - workspaces
    post:
      consumes:
      - application/json
      description: Add a member to a workspace
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Member info
        in: body
        name: member
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add workspace member
      tags:
      - workspaces
  /workspaces/{id}/members/{user_id}:
    delete:
      description: Remove a member from a workspace
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Member ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove workspace member
      tags:
      - workspaces
...
  /workspaces/me:
    get:
      description: Get all workspaces where the authenticated user has membership
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Workspace'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get my workspaces
      tags:
      - workspaces
swagger: "2.0"